test: switch from jest to vitest
