--- jest.preset.js
@@ -1,3 +1,0 @@
-const nxPreset = require('@nrwl/jest/preset');
-
-module.exports = { ...nxPreset };
