+++ package.json
@@ -36,24 +36,21 @@
         "precommit": "lint-staged",
         "prepack": "npx rimraf dist && tsc",
         "semantic-release": "semantic-release",
-        "test": "dotenv -e .env -- cross-var jest",
+        "test": "dotenv -e .env -- cross-var vitest",
         "test-ci": "jest",
-        "test-watch": "dotenv -e .env -- cross-var jest --watch"
+        "test-watch": "dotenv -e .env -- cross-var vitest --watch"
     },
     "devDependencies": {
         "@commitlint/cli": "^16.2.1",
         "@commitlint/config-angular": "^16.2.1",
         "@nrwl/eslint-plugin-nx": "13.7.2",
-        "@nrwl/jest": "13.7.2",
         "@nrwl/react": "13.7.2",
         "@semantic-release/changelog": "^6.0.1",
         "@semantic-release/git": "^10.0.1",
         "@types/http-proxy": "^1.17.5",
-        "@types/jest": "^27.4.1",
         "@types/node": "^14.0.13",
         "@typescript-eslint/eslint-plugin": "^5.13.0",
         "@typescript-eslint/parser": "^5.13.0",
-        "babel-jest": "27.2.3",
         "cross-var": "^1.1.0",
         "dotenv-cli": "^4.0.0",
         "eslint": "^7.32.0",
@@ -61,16 +58,16 @@
         "eslint-plugin-tsdoc": "^0.2.14",
         "esm": "^3.2.25",
         "husky": "^7.0.4",
-        "jest": "27.2.3",
         "lint-staged": "^10.2.9",
         "prettier": "^2.5.1",
         "rimraf": "^3.0.2",
         "semantic-release": "^19.0.2",
         "shelljs": "^0.8.4",
-        "ts-jest": "27.0.5",
         "typedoc": "^0.22.12",
         "typedoc-plugin-markdown": "^3.11.14",
-        "typescript": "4.3.5"
+        "typescript": "4.3.5",
+        "vite-tsconfig-paths": "^3.4.1",
+        "vitest": "^0.12.4"
     },
     "dependencies": {
         "@fauna-labs/fauna-schema-migrate": "^2.2.1",
