+++ tests/constants.js
@@ -1,2 +1,2 @@
 export const FAUNA_TEST_TIMEOUT = 60 * 1000 * 2; // two minutes
-export const EMAIL_TEST_TIMEOUT = 1000; // one second
+export const SHORT_TEST_TIMEOUT = 1000; // one second
