+++ tests/fauna/access.test.ts
@@ -6,7 +6,6 @@ import {
     setupTestDatabase,
     populateDatabaseSchemaFromFiles,
 } from './helpers/_setup-db';
-import { FAUNA_TEST_TIMEOUT } from '../constants';
 import type {
     TestContext,
     FaunaLoginResult,
@@ -17,8 +16,6 @@ import type {
 
 const q = fauna.query;
 const { Call, Create, Collection, Get, Paginate, Documents } = q;
-
-jest.setTimeout(FAUNA_TEST_TIMEOUT);
 
 const setUp: SetUp = async testName => {
     const context: TestContext = {
