+++ tests/fauna/loginWithMagicLink.test.ts
@@ -5,7 +5,6 @@ import {
     setupTestDatabase,
     populateDatabaseSchemaFromFiles,
 } from './helpers/_setup-db';
-import { FAUNA_TEST_TIMEOUT } from '../constants';
 import type {
     TestContext,
     FaunaLoginResult,
@@ -17,8 +16,6 @@ import type {
 
 const q = fauna.query;
 const { Call, Paginate, Lambda, Get, Var, Tokens } = q;
-
-jest.setTimeout(FAUNA_TEST_TIMEOUT);
 
 const setUp: SetUp = async testName => {
     const context: TestContext = {
