+++ tests/fauna/refresh.test.ts
@@ -17,7 +17,6 @@ import {
     REFRESH_TOKEN_RECLAIMTIME_SECONDS,
 } from './resources/functions/_refresh-modified';
 import { verifyTokens } from './helpers/_test-extensions';
-import { FAUNA_TEST_TIMEOUT } from '../constants';
 import type {
     TestContext,
     FaunaLoginResult,
@@ -41,8 +40,6 @@ const {
     Lambda,
     Var,
 } = q;
-
-jest.setTimeout(FAUNA_TEST_TIMEOUT);
 
 const setUp: SetUp = async testName => {
     const context: TestContext = {
