+++ tests/fauna/sample-auth-flow.test.ts
@@ -4,7 +4,6 @@ import {
     setupTestDatabase,
     populateDatabaseSchemaFromFiles,
 } from './helpers/_setup-db';
-import { FAUNA_TEST_TIMEOUT } from '../constants';
 import type {
     TestContext,
     CreateKeyResult,
@@ -16,8 +15,6 @@ import type {
 
 const q = fauna.query;
 const { Call, CreateKey, Role } = q;
-
-jest.setTimeout(FAUNA_TEST_TIMEOUT);
 
 const clientDomain = process.env.FAUNADB_DOMAIN
     ? process.env.FAUNADB_DOMAIN
